requirements.txt
# Core dependencies
python-dotenv==1.1.1
requests==2.32.5
pydantic==2.11.7
tiktoken==0.8.0

# Azure services
azure-identity==1.24.0
azure-cosmos==4.9.0
azure-search-documents==11.5.3
azure-ai-generative==1.0.0b11
azure-ai-projects>=1.0.0b1
packaging>=23.2,<25

# Semantic Kernel
semantic-kernel==1.36.1

# Optional: For better performance
numpy==2.3.2

# Testing
pytest==8.0.0
pytest-mock==3.12.0